native-port: fix ILP32 pointer-truncation in JKRExpHeap + J3DTexture (uintptr_t)
For the architecture-independent native PC build (x86-64/arm64): the decomp casts
pointers to u32 (ILP32 assumption) which truncates under LP64. Use uintptr_t for the
pointer arithmetic so it's correct on any 64-bit arch. First slice of the campaign.

+++ include/JSystem/J3D/J3DGraphBase/J3DTexture.hpp
@@ -21,11 +21,11 @@ public:
 	{
 		mResources[entry] = timg;
 		mResources[entry].imageDataOffset
-		    = ((mResources[entry].imageDataOffset + (u32)&timg
-		        - (u32)(mResources + entry)));
+		    = ((mResources[entry].imageDataOffset + (uintptr_t)&timg
+		        - (uintptr_t)(mResources + entry)));
 		mResources[entry].paletteOffset
-		    = ((mResources[entry].paletteOffset + (u32)&timg
-		        - (u32)(mResources + entry)));
+		    = ((mResources[entry].paletteOffset + (uintptr_t)&timg
+		        - (uintptr_t)(mResources + entry)));
 	}
 };
